BugFix: Fixed overflow on frame-index
changed from uint8 to uint16 to allow sequences of >255 bytes

+++ EmbeddedSoftware/SandBox/LowLevel/I2C/I2C.h
@@ -19,7 +19,7 @@ public:
 protected:
 private:
 	static uint8_t *buffer_ref;
-	static uint8_t tx_index;
+	static uint16_t tx_index;
 	static uint16_t tx_count;
 
 	static volatile bool tx_complete;
